Add Slop dependency.
Begin to use Slop to parse arguments after the initial two arguments are
gathered.

Also raise an error if something has gone wrong with passing
in a value for location. This can happen if just a flag is passed in for
now, but this needs to be changed in the future to allow for just one
flag such as help and version.